refactor: extract indent rule handlers into dedicated module
