+++ packages/eslint-plugin/rules/indent/indent.ts
@@ -4,11 +4,28 @@
  * This is done intentionally based on the internal implementation of the base indent rule.
  */
 
-import type { ASTNode, JSONSchema, NodeTypes, RuleFunction, RuleListener, Token, Tree } from '#types'
+import type { ASTNode, JSONSchema, RuleFunction, RuleListener, Token, Tree } from '#types'
 import type { MessageIds, RuleOptions } from './types'
-import { AST_NODE_TYPES, createGlobalLinebreakMatcher, getCommentsBetween, isClosingBraceToken, isClosingBracketToken, isClosingParenToken, isColonToken, isCommentToken, isEqToken, isNotClosingParenToken, isNotOpeningParenToken, isNotSemicolonToken, isOpeningBraceToken, isOpeningBracketToken, isOpeningParenToken, isOptionalChainPunctuator, isQuestionToken, isSemicolonToken, isSingleLine, isTokenOnSameLine, skipChainExpression, STATEMENT_LIST_PARENTS } from '#utils/ast'
+import type { IndentContext } from './utils/handlers'
+import { AST_NODE_TYPES, getCommentsBetween, isClosingBraceToken, isClosingParenToken, isColonToken, isCommentToken, isEqToken, isNotOpeningParenToken, isNotSemicolonToken, isOpeningBraceToken, isOpeningParenToken, isSemicolonToken, isSingleLine, isTokenOnSameLine } from '#utils/ast'
 import { createRule } from '#utils/create-rule'
 import { isESTreeSourceCode } from '#utils/eslint-core'
+import {
+  addBlocklessNodeIndent,
+  addElementListIndent,
+  addFunctionCallIndent,
+  addParensIndent,
+  checkArrayLikeNode,
+  checkAssignmentOperator,
+  checkBlockLikeNode,
+  checkClassProperty,
+  checkConditionalNode,
+  checkHeritages,
+  checkMemberExpression,
+  checkObjectLikeNode,
+  checkOperatorToken,
+  ignoreNode,
+} from './utils/handlers'
 import { OffsetStorage } from './utils/offset-storage'
 import { TokenInfo } from './utils/token-info'
 import { hasBlankLinesBetween, validateTokenIndent } from './utils/validator'
@@ -167,8 +184,6 @@ const KNOWN_NODES = new Set([
   // AST_NODE_TYPES.TSIntersectionType,
   // AST_NODE_TYPES.TSUnionType,
 ])
-
-type ElementListOffset = 'first' | 'off' | number
 
 /*
  * General rule strategy:
@@ -395,7 +410,7 @@ export default createRule<RuleOptions, MessageIds>({
     const DEFAULT_FUNCTION_BODY_INDENT = 1
     const DEFAULT_FUNCTION_RETURN_TYPE_INDENT = 1
 
-    let indentType = 'space'
+    let indentType: IndentContext['indentType'] = 'space'
     let indentSize = 4
     const options = {
       SwitchCase: 0,
@@ -464,6 +479,26 @@ export default createRule<RuleOptions, MessageIds>({
     const offsets = new OffsetStorage(tokenInfo, indentSize, indentType === 'space' ? ' ' : '\t', sourceCode.text.length)
     const parameterParens = new WeakSet()
 
+    const ternaryOptions: false | Partial<Record<string, boolean>> = options.offsetTernaryExpressions !== false
+      ? {
+          CallExpression: true,
+          AwaitExpression: true,
+          NewExpression: true,
+          ...options.offsetTernaryExpressions === true ? {} : options.offsetTernaryExpressions,
+        }
+      : false
+
+    const ctx: IndentContext = {
+      sourceCode,
+      offsets,
+      tokenInfo,
+      parameterParens,
+      indentSize,
+      indentType,
+      options,
+      ternaryOptions,
+    }
+
     /**
      * Creates an error message for a line, given the expected/actual indentation.
      * @param expectedAmount The expected amount of indentation characters for this line
@@ -523,541 +558,19 @@ export default createRule<RuleOptions, MessageIds>({
       })
     }
 
-    /**
-     * Check to see if the node is a file level IIFE
-     * @param node The function node to check.
-     * @returns True if the node is the outer IIFE
-     */
-    function isOuterIIFE(node: ASTNode) {
-      /**
-       * Verify that the node is an IIFE
-       */
-      if (!node.parent || node.parent.type !== 'CallExpression' || node.parent.callee !== node)
-        return false
-
-      /**
-       * Navigate legal ancestors to determine whether this IIFE is outer.
-       * A "legal ancestor" is an expression or statement that causes the function to get executed immediately.
-       * For example, `!(function(){})()` is an outer IIFE even though it is preceded by a ! operator.
-       */
-      let statement = node.parent && node.parent.parent
-
-      while (
-        statement.type === 'UnaryExpression' && ['!', '~', '+', '-'].includes(statement.operator)
-        || statement.type === 'AssignmentExpression'
-        || statement.type === 'LogicalExpression'
-        || statement.type === 'SequenceExpression'
-        || statement.type === 'VariableDeclarator'
-      ) {
-        statement = statement.parent
-      }
-
-      return (statement.type === 'ExpressionStatement' || statement.type === 'VariableDeclaration') && statement.parent.type === 'Program'
-    }
-
-    /**
-     * Counts the number of linebreaks that follow the last non-whitespace character in a string
-     * @param string The string to check
-     * @returns The number of JavaScript linebreaks that follow the last non-whitespace character,
-     * or the total number of linebreaks if the string is all whitespace.
-     */
-    function countTrailingLinebreaks(string: string) {
-      const trailingWhitespace = string.match(/\s*$/u)![0]
-      const linebreakMatches = trailingWhitespace.match(createGlobalLinebreakMatcher())
-
-      return linebreakMatches === null ? 0 : linebreakMatches.length
-    }
-
-    /**
-     * Check indentation for lists of elements (arrays, objects, function params)
-     * @param elements List of elements that should be offset
-     * @param startToken The start token of the list that element should be aligned against, e.g. '['
-     * @param endToken The end token of the list, e.g. ']'
-     * @param offset The amount that the elements should be offset
-     */
-    function addElementListIndent(elements: (ASTNode | null)[], startToken: Token, endToken: Token, offset: ElementListOffset) {
-      if (isTokenOnSameLine(startToken, endToken))
-        return
-
-      /**
-       * Gets the first token of a given element, including surrounding parentheses.
-       * @param element A node in the `elements` list
-       * @returns The first token of this element
-       */
-      function getFirstToken(element: ASTNode) {
-        let token: Token = sourceCode.getTokenBefore(element)!
-
-        while (isOpeningParenToken(token) && token !== startToken)
-          token = sourceCode.getTokenBefore(token)!
-
-        return sourceCode.getTokenAfter(token)!
-      }
-
-      // Run through all the tokens in the list, and offset them by one indent level (mainly for comments, other things will end up overridden)
-      offsets.setDesiredOffsets(
-        [startToken.range[1], endToken.range[0]],
-        startToken,
-        typeof offset === 'number' ? offset : 1,
-      )
-      offsets.setDesiredOffset(endToken, startToken, 0)
-
-      // If the preference is "first" but there is no first element (e.g. sparse arrays w/ empty first slot), fall back to 1 level.
-      if (offset === 'first' && elements.length && !elements[0])
-        return
-
-      elements
-        .forEach((element, index) => {
-          if (!element) {
-            // Skip holes in arrays
-            return
-          }
-          if (offset === 'off') {
-            // Ignore the first token of every element if the "off" option is used
-            offsets.ignoreToken(getFirstToken(element))
-          }
-
-          // Offset the following elements correctly relative to the first element
-          if (index === 0)
-            return
-
-          if (offset === 'first' && tokenInfo.isFirstTokenOfLine(getFirstToken(element))) {
-            offsets.matchOffsetOf(getFirstToken(elements[0]!), getFirstToken(element))
-          }
-          else {
-            const previousElement = elements[index - 1]!
-            const firstTokenOfPreviousElement = previousElement && getFirstToken(previousElement)!
-            const previousElementLastToken = previousElement && sourceCode.getLastToken(previousElement)!
-
-            if (
-              previousElement && previousElementLastToken.loc.end.line - countTrailingLinebreaks(previousElementLastToken.value) > startToken.loc.end.line
-            ) {
-              offsets.setDesiredOffsets(
-                [previousElement.range[1], element.range[1]],
-                firstTokenOfPreviousElement,
-                0,
-              )
-            }
-          }
-        })
-    }
-
-    /**
-     * Check and decide whether to check for indentation for blockless nodes
-     * Scenarios are for or while statements without braces around them
-     * @param node node to examine
-     */
-    function addBlocklessNodeIndent(node: ASTNode) {
-      if (node.type !== 'BlockStatement') {
-        const lastParentToken = sourceCode.getTokenBefore(node, isNotOpeningParenToken)!
-
-        let firstBodyToken = sourceCode.getFirstToken(node)!
-        let lastBodyToken = sourceCode.getLastToken(node)!
-
-        while (
-          isOpeningParenToken(sourceCode.getTokenBefore(firstBodyToken)!)
-          && isClosingParenToken(sourceCode.getTokenAfter(lastBodyToken)!)
-        ) {
-          firstBodyToken = sourceCode.getTokenBefore(firstBodyToken)!
-          lastBodyToken = sourceCode.getTokenAfter(lastBodyToken)!
-        }
-
-        offsets.setDesiredOffsets([firstBodyToken.range[0], lastBodyToken.range[1]], lastParentToken, 1)
-      }
-    }
-
-    /**
-     * Checks the indentation for nodes that are like function calls (`CallExpression` and `NewExpression`)
-     * @param node A CallExpression or NewExpression node
-     */
-    function addFunctionCallIndent(node: Tree.CallExpression | Tree.NewExpression) {
-      let openingParen
-
-      if (node.arguments.length) {
-        openingParen = sourceCode.getTokenAfter(
-          node.typeArguments ?? node.callee,
-          isOpeningParenToken,
-        )!
-      }
-      else {
-        openingParen = sourceCode.getLastToken(node, 1)!
-      }
-
-      const closingParen = sourceCode.getLastToken(node)!
-
-      parameterParens.add(openingParen)
-      parameterParens.add(closingParen)
-
-      /**
-       * If `?.` token exists, set desired offset for that.
-       * This logic is copied from `MemberExpression`'s.
-       */
-      if ('optional' in node && node.optional) {
-        const dotToken = sourceCode.getTokenAfter(node.callee, isOptionalChainPunctuator)!
-        const calleeParenCount = sourceCode.getTokensBetween(node.callee, dotToken, { filter: isClosingParenToken }).length
-        const firstTokenOfCallee = calleeParenCount
-          ? sourceCode.getTokenBefore(node.callee, { skip: calleeParenCount - 1 })!
-          : sourceCode.getFirstToken(node.callee)!
-        const lastTokenOfCallee = sourceCode.getTokenBefore(dotToken)!
-        const offsetBase = isTokenOnSameLine(lastTokenOfCallee, openingParen)
-          ? lastTokenOfCallee
-          : firstTokenOfCallee
-
-        offsets.setDesiredOffset(dotToken, offsetBase, 1)
-      }
-
-      const offsetAfterToken = node.callee.type === 'TaggedTemplateExpression'
-        ? sourceCode.getFirstToken(node.callee.quasi)!
-        : node.typeArguments ?? openingParen
-      const offsetToken = sourceCode.getTokenBefore(offsetAfterToken)!
-
-      offsets.setDesiredOffset(openingParen, offsetToken, 0)
-
-      addElementListIndent(node.arguments, openingParen, closingParen, options.CallExpression.arguments)
-    }
-
-    /**
-     * Checks the indentation of parenthesized values, given a list of tokens in a program
-     * @param tokens A list of tokens
-     */
-    function addParensIndent(tokens: Token[]) {
-      const parenStack = []
-      const parenPairs = []
-
-      for (let i = 0; i < tokens.length; i++) {
-        const nextToken = tokens[i]
-
-        if (isOpeningParenToken(nextToken))
-          parenStack.push(nextToken)
-        else if (isClosingParenToken(nextToken))
-          parenPairs.push({ left: parenStack.pop(), right: nextToken })
-      }
-
-      for (let i = parenPairs.length - 1; i >= 0; i--) {
-        const leftParen = parenPairs[i].left!
-        const rightParen = parenPairs[i].right!
-
-        if (isTokenOnSameLine(leftParen, rightParen))
-          continue
-
-        // We only want to handle parens around expressions, so exclude parentheses that are in function parameters and function call arguments.
-        if (!parameterParens.has(leftParen) && !parameterParens.has(rightParen)) {
-          const parenthesizedTokens: Set<Token> = new Set(sourceCode.getTokensBetween(leftParen, rightParen))
-
-          parenthesizedTokens.forEach((token) => {
-            const dependency = offsets.getFirstDependency(token)
-
-            if (!dependency || !parenthesizedTokens.has(dependency))
-              offsets.setDesiredOffset(token, leftParen, 1)
-          })
-        }
-
-        offsets.setDesiredOffset(rightParen, leftParen, 0)
-      }
-    }
-
-    /**
-     * Ignore all tokens within an unknown node whose offset do not depend
-     * on another token's offset within the unknown node
-     * @param node Unknown Node
-     */
-    function ignoreNode(node: ASTNode) {
-      const unknownNodeTokens: Set<Token> = new Set(sourceCode.getTokens(node, { includeComments: true }))
-
-      unknownNodeTokens.forEach((token) => {
-        const dependency = offsets.getFirstDependency(token)
-
-        if (!dependency || !unknownNodeTokens.has(dependency)) {
-          const firstTokenOfLine = tokenInfo.getFirstTokenOfLine(token)!
-
-          if (token === firstTokenOfLine)
-            offsets.ignoreToken(token)
-          else
-            offsets.setDesiredOffset(token, firstTokenOfLine, 0)
-        }
-      })
-    }
-
-    /**
-     * Check whether the given token is on the first line of a statement.
-     * @param token The token to check.
-     * @param leafNode The expression node that the token belongs directly.
-     * @returns `true` if the token is on the first line of a statement.
-     */
-    function isOnFirstLineOfStatement(token: Token, leafNode: ASTNode): boolean {
-      let node = leafNode
-
-      while (node.parent && !node.parent.type.endsWith('Statement') && !node.parent.type.endsWith('Declaration'))
-        node = node.parent
-
-      node = node.parent!
-
-      return !node || node.loc.start.line === token.loc.start.line
-    }
-
     const ignoredNodeFirstTokens = new Set<Token>()
 
-    function checkAssignmentOperator(operator: Token) {
-      const left = sourceCode.getTokenBefore(operator)!
-      const right = sourceCode.getTokenAfter(operator)!
-
-      if (typeof options.assignmentOperator === 'number') {
-        offsets.setDesiredOffset(operator, left, options.assignmentOperator)
-        offsets.setDesiredOffset(right, operator, options.assignmentOperator)
-      }
-      else {
-        offsets.ignoreToken(operator)
-        offsets.ignoreToken(right)
-      }
-    }
-
-    function checkArrayLikeNode(node: Tree.ArrayExpression | Tree.ArrayPattern | Tree.TSTupleType) {
-      const elementList = node.type === AST_NODE_TYPES.TSTupleType ? node.elementTypes : node.elements
-      const openingBracket = sourceCode.getFirstToken(node)!
-      const closingBracket = sourceCode.getTokenAfter([...elementList].reverse().find(_ => _) || openingBracket, isClosingBracketToken)!
-
-      addElementListIndent(elementList, openingBracket, closingBracket, options.ArrayExpression)
-    }
-
-    function checkObjectLikeNode(node: Tree.ObjectExpression | Tree.ObjectPattern | Tree.TSEnumDeclaration | Tree.TSTypeLiteral, properties: ASTNode[]) {
-      const openingCurly = sourceCode.getFirstToken(node, isOpeningBraceToken)!
-      const closingCurly = sourceCode.getTokenAfter(
-        properties.length ? properties[properties.length - 1] : openingCurly,
-        isClosingBraceToken,
-      )!
-
-      addElementListIndent(properties, openingCurly, closingCurly, options.ObjectExpression)
-    }
-
-    const ternaryOptions: false | Partial<Record<NodeTypes, boolean>> = options.offsetTernaryExpressions !== false
-      ? {
-          CallExpression: true,
-          AwaitExpression: true,
-          NewExpression: true,
-          ...options.offsetTernaryExpressions === true ? {} : options.offsetTernaryExpressions,
-        }
-      : false
-
-    function checkConditionalNode(node: Tree.ConditionalExpression | Tree.TSConditionalType, test: ASTNode, consequent: ASTNode, alternate: ASTNode) {
-      const firstToken = sourceCode.getFirstToken(node)!
-
-      // `flatTernaryExpressions` option is for the following style:
-      // var a =
-      //     foo > 0 ? bar :
-      //     foo < 0 ? baz :
-      //     /*else*/ qiz ;
-      if (options.flatTernaryExpressions && isTokenOnSameLine(test, consequent) && !isOnFirstLineOfStatement(firstToken, node))
-        return
-
-      function checkBranch(branch: ASTNode, branchFirstToken: Token) {
-        let offset = 1
-        if (ternaryOptions) {
-          const branchType = skipChainExpression(branch).type
-
-          if (branchFirstToken.type === 'Punctuator' || ternaryOptions[branchType]) {
-            offset = 2
-          }
-        }
-
-        offsets.setDesiredOffset(
-          branchFirstToken,
-          firstToken,
-          offset,
-        )
-      }
-
-      const questionMarkToken = sourceCode.getFirstTokenBetween(test, consequent, isQuestionToken)!
-      const colonToken = sourceCode.getFirstTokenBetween(consequent, alternate, isColonToken)!
-
-      const firstConsequentToken = sourceCode.getTokenAfter(questionMarkToken)!
-      const lastConsequentToken = sourceCode.getTokenBefore(colonToken)!
-      const firstAlternateToken = sourceCode.getTokenAfter(colonToken)!
-
-      offsets.setDesiredOffset(questionMarkToken, firstToken, 1)
-      offsets.setDesiredOffset(colonToken, firstToken, 1)
-
-      checkBranch(consequent, firstConsequentToken)
-
-      /**
-       * The alternate and the consequent should usually have the same indentation.
-       * If they share part of a line, align the alternate against the first token of the consequent.
-       * This allows the alternate to be indented correctly in cases like this:
-       * foo ? (
-       *   bar
-       * ) : ( // this '(' is aligned with the '(' above, so it's considered to be aligned with `foo`
-       *   baz // as a result, `baz` is offset by 1 rather than 2
-       * )
-       */
-      if (isTokenOnSameLine(lastConsequentToken, firstAlternateToken)) {
-        offsets.setDesiredOffset(firstAlternateToken, firstConsequentToken, 0)
-      }
-      else {
-        /**
-         * If the alternate and consequent do not share part of a line, offset the alternate from the first
-         * token of the conditional expression. For example:
-         * foo ? bar
-         *   : baz
-         *
-         * If `baz` were aligned with `bar` rather than being offset by 1 from `foo`, `baz` would end up
-         * having no expected indentation.
-         */
-        checkBranch(alternate, firstAlternateToken)
-      }
-    }
-
-    function checkOperatorToken(left: ASTNode, right: ASTNode, operator: string) {
-      const operatorToken = sourceCode.getFirstTokenBetween(left, right, token => token.value === operator)!
-
-      /**
-       * For backwards compatibility, don't check BinaryExpression indents, e.g.
-       * var foo = bar &&
-       *                   baz;
-       */
-
-      const tokenAfterOperator = sourceCode.getTokenAfter(operatorToken)!
-      offsets.ignoreToken(operatorToken)
-      offsets.ignoreToken(tokenAfterOperator)
-      offsets.setDesiredOffset(tokenAfterOperator, operatorToken, 0)
-    }
-
-    function checkMemberExpression(
-      node: Tree.MemberExpression | Tree.JSXMemberExpression | Tree.MetaProperty | Tree.TSIndexedAccessType | Tree.TSQualifiedName,
-      object: ASTNode,
-      property: ASTNode,
-      computed = false,
-    ) {
-      const firstNonObjectToken = sourceCode.getFirstTokenBetween(object, property, isNotClosingParenToken)!
-      const secondNonObjectToken = sourceCode.getTokenAfter(firstNonObjectToken)!
-
-      const objectParenCount = sourceCode.getTokensBetween(object, property, { filter: isClosingParenToken }).length
-      const firstObjectToken = objectParenCount
-        ? sourceCode.getTokenBefore(object, { skip: objectParenCount - 1 })!
-        : sourceCode.getFirstToken(object)!
-      const lastObjectToken = sourceCode.getTokenBefore(firstNonObjectToken)!
-      const firstPropertyToken = computed ? firstNonObjectToken : secondNonObjectToken
-
-      if (computed) {
-        // For computed MemberExpressions, match the closing bracket with the opening bracket.
-        offsets.setDesiredOffset(sourceCode.getLastToken(node)!, firstNonObjectToken, 0)
-        offsets.setDesiredOffsets(property.range, firstNonObjectToken, 1)
-      }
-
-      /**
-       * If the object ends on the same line that the property starts, match against the last token
-       * of the object, to ensure that the MemberExpression is not indented.
-       *
-       * Otherwise, match against the first token of the object, e.g.
-       * foo
-       *   .bar
-       *   .baz // <-- offset by 1 from `foo`
-       */
-      const offsetBase = isTokenOnSameLine(lastObjectToken, firstPropertyToken)
-        ? lastObjectToken
-        : firstObjectToken
-
-      if (typeof options.MemberExpression === 'number') {
-        // Match the dot (for non-computed properties) or the opening bracket (for computed properties) against the object.
-        offsets.setDesiredOffset(firstNonObjectToken, offsetBase, options.MemberExpression)
-
-        /**
-         * For computed MemberExpressions, match the first token of the property against the opening bracket.
-         * Otherwise, match the first token of the property against the object.
-         */
-        offsets.setDesiredOffset(secondNonObjectToken, computed ? firstNonObjectToken : offsetBase, options.MemberExpression)
-      }
-      else {
-        // If the MemberExpression option is off, ignore the dot and the first token of the property.
-        offsets.ignoreToken(firstNonObjectToken)
-        offsets.ignoreToken(secondNonObjectToken)
-
-        // To ignore the property indentation, ensure that the property tokens depend on the ignored tokens.
-        offsets.setDesiredOffset(firstNonObjectToken, offsetBase, 0)
-        offsets.setDesiredOffset(secondNonObjectToken, firstNonObjectToken, 0)
-      }
-    }
-
-    function checkBlockLikeNode(node: Tree.BlockStatement | Tree.ClassBody | Tree.TSInterfaceBody | Tree.TSEnumBody | Tree.TSModuleBlock) {
-      let blockIndentLevel
-
-      if (node.parent && isOuterIIFE(node.parent))
-        blockIndentLevel = options.outerIIFEBody
-      else if (node.parent && (node.parent.type === 'FunctionExpression' || node.parent.type === 'ArrowFunctionExpression'))
-        blockIndentLevel = options.FunctionExpression.body
-      else if (node.parent && node.parent.type === 'FunctionDeclaration')
-        blockIndentLevel = options.FunctionDeclaration.body
-      else
-        blockIndentLevel = 1
-
-      /**
-       * For blocks that aren't lone statements, ensure that the opening curly brace
-       * is aligned with the parent.
-       */
-      if (!STATEMENT_LIST_PARENTS.has(node.parent.type))
-        offsets.setDesiredOffset(sourceCode.getFirstToken(node)!, sourceCode.getFirstToken(node.parent)!, 0)
-
-      addElementListIndent(
-        node.type === AST_NODE_TYPES.TSEnumBody ? node.members : node.body,
-        sourceCode.getFirstToken(node)!,
-        sourceCode.getLastToken(node)!,
-        blockIndentLevel,
-      )
-    }
-
-    function checkHeritages(node: Tree.ClassDeclaration | Tree.ClassExpression | Tree.TSInterfaceDeclaration, heritages: ASTNode[]) {
-      const classToken = sourceCode.getFirstToken(node)!
-      const extendsToken = sourceCode.getTokenBefore(heritages[0], isNotOpeningParenToken)!
-
-      offsets.setDesiredOffsets([extendsToken.range[0], node.body.range[0]], classToken, 1)
-    }
-
-    function checkClassProperty(node: Tree.PropertyDefinition | Tree.AccessorProperty | Tree.TSAbstractPropertyDefinition | Tree.TSAbstractAccessorProperty) {
-      const firstToken = sourceCode.getFirstToken(node)!
-      const lastToken = sourceCode.getLastToken(node)!
-      let keyLastToken: Token
-
-      // Indent key.
-      if (node.computed) {
-        const bracketTokenL = sourceCode.getTokenBefore(node.key, isOpeningBracketToken)!
-        const bracketTokenR = keyLastToken = sourceCode.getTokenAfter(node.key, isClosingBracketToken)!
-        const keyRange: [number, number] = [bracketTokenL.range[1], bracketTokenR.range[0]]
-
-        if (bracketTokenL !== firstToken)
-          offsets.setDesiredOffset(bracketTokenL, firstToken, 0)
-
-        offsets.setDesiredOffsets(keyRange, bracketTokenL, 1)
-        offsets.setDesiredOffset(bracketTokenR, bracketTokenL, 0)
-      }
-      else {
-        const idToken = keyLastToken = sourceCode.getFirstToken(node.key)!
-
-        if (!node.decorators?.length && idToken !== firstToken)
-          offsets.setDesiredOffset(idToken, firstToken, 1)
-      }
-
-      // Indent initializer.
-      if (node.value) {
-        const operator = sourceCode.getTokenBefore(node.value, isEqToken)!
-        checkAssignmentOperator(operator)
-
-        if (isSemicolonToken(lastToken))
-          offsets.setDesiredOffset(lastToken, operator, 1)
-      }
-      else if (isSemicolonToken(lastToken)) {
-        // TODO: ignore like `VariableDeclaration`
-        offsets.setDesiredOffset(lastToken, keyLastToken, 1)
-      }
-    }
-
     const baseOffsetListeners: RuleListener = {
-      'ArrayExpression': checkArrayLikeNode,
+      'ArrayExpression': node => checkArrayLikeNode(ctx, node),
 
-      'ArrayPattern': checkArrayLikeNode,
+      'ArrayPattern': node => checkArrayLikeNode(ctx, node),
 
       ObjectExpression(node) {
-        checkObjectLikeNode(node, node.properties)
+        checkObjectLikeNode(ctx, node, node.properties)
       },
 
       ObjectPattern(node) {
-        checkObjectLikeNode(node, node.properties)
+        checkObjectLikeNode(ctx, node, node.properties)
       },
 
       ArrowFunctionExpression(node) {
@@ -1072,54 +585,54 @@ export default createRule<RuleOptions, MessageIds>({
 
           parameterParens.add(openingParen)
           parameterParens.add(closingParen)
-          addElementListIndent(node.params, openingParen, closingParen, options.FunctionExpression.parameters)
+          addElementListIndent(ctx, node.params, openingParen, closingParen, options.FunctionExpression.parameters)
         }
 
-        addBlocklessNodeIndent(node.body)
+        addBlocklessNodeIndent(ctx, node.body)
       },
 
       AssignmentExpression(node) {
         const operator = sourceCode.getFirstTokenBetween(node.left, node.right, token => token.value === node.operator)!
 
-        checkAssignmentOperator(operator)
+        checkAssignmentOperator(ctx, operator)
       },
 
       AssignmentPattern(node) {
         const operator = sourceCode.getFirstTokenBetween(node.left, node.right, isEqToken)!
 
-        checkAssignmentOperator(operator)
+        checkAssignmentOperator(ctx, operator)
       },
 
       BinaryExpression(node) {
-        checkOperatorToken(node.left, node.right, node.operator)
+        checkOperatorToken(ctx, node.left, node.right, node.operator)
       },
 
       LogicalExpression(node) {
-        checkOperatorToken(node.left, node.right, node.operator)
+        checkOperatorToken(ctx, node.left, node.right, node.operator)
       },
 
-      'BlockStatement': checkBlockLikeNode,
+      'BlockStatement': node => checkBlockLikeNode(ctx, node),
 
-      'ClassBody': checkBlockLikeNode,
+      'ClassBody': node => checkBlockLikeNode(ctx, node),
 
-      'CallExpression': addFunctionCallIndent,
+      'CallExpression': node => addFunctionCallIndent(ctx, node),
 
       ClassDeclaration(node) {
         if (!node.superClass)
           return
 
-        checkHeritages(node, [node.superClass])
+        checkHeritages(ctx, node, [node.superClass])
       },
 
       ClassExpression(node) {
         if (!node.superClass)
           return
 
-        checkHeritages(node, [node.superClass])
+        checkHeritages(ctx, node, [node.superClass])
       },
 
       ConditionalExpression(node) {
-        checkConditionalNode(node, node.test, node.consequent, node.alternate)
+        checkConditionalNode(ctx, node, node.test, node.consequent, node.alternate)
       },
 
       'DoWhileStatement, WhileStatement, ForInStatement, ForOfStatement, WithStatement': function (
@@ -1130,7 +643,7 @@ export default createRule<RuleOptions, MessageIds>({
           | Tree.ForOfStatement
           | Tree.WithStatement,
       ) {
-        addBlocklessNodeIndent(node.body)
+        addBlocklessNodeIndent(ctx, node.body)
       },
 
       ExportNamedDeclaration(node) {
@@ -1140,7 +653,7 @@ export default createRule<RuleOptions, MessageIds>({
             : sourceCode.getLastToken(node, isClosingBraceToken)!
 
           // Indent the specifiers in `export {foo, bar, baz}`
-          addElementListIndent(node.specifiers, sourceCode.getFirstToken(node, { skip: 1 })!, closingCurly, 1)
+          addElementListIndent(ctx, node.specifiers, sourceCode.getFirstToken(node, { skip: 1 })!, closingCurly, 1)
 
           if (node.source) {
             const fromToken = sourceCode.getTokenAfter(
@@ -1156,7 +669,7 @@ export default createRule<RuleOptions, MessageIds>({
               // Has attributes
               const openingCurly = sourceCode.getTokenAfter(node.source, isOpeningBraceToken)!
               const closingCurly = lastClosingCurly
-              addElementListIndent(node.attributes, openingCurly, closingCurly, 1)
+              addElementListIndent(ctx, node.attributes, openingCurly, closingCurly, 1)
             }
           }
         }
@@ -1177,7 +690,7 @@ export default createRule<RuleOptions, MessageIds>({
           // Has attributes
           const openingCurly = sourceCode.getTokenAfter(node.source, isOpeningBraceToken)!
           const closingCurly = lastClosingCurly
-          addElementListIndent(node.attributes, openingCurly, closingCurly, 1)
+          addElementListIndent(ctx, node.attributes, openingCurly, closingCurly, 1)
         }
       },
 
@@ -1193,7 +706,7 @@ export default createRule<RuleOptions, MessageIds>({
         if (node.update)
           offsets.setDesiredOffsets(node.update.range, forOpeningParen, 1)
 
-        addBlocklessNodeIndent(node.body)
+        addBlocklessNodeIndent(ctx, node.body)
       },
 
       'FunctionDeclaration, FunctionExpression': function (node: Tree.FunctionDeclaration | Tree.FunctionExpression) {
@@ -1211,7 +724,7 @@ export default createRule<RuleOptions, MessageIds>({
 
         parameterParens.add(paramsOpeningParen)
         parameterParens.add(paramsClosingParen)
-        addElementListIndent(node.params, paramsOpeningParen, paramsClosingParen, options[node.type].parameters)
+        addElementListIndent(ctx, node.params, paramsOpeningParen, paramsClosingParen, options[node.type].parameters)
 
         if (node.returnType) {
           offsets.setDesiredOffsets(node.returnType.range, paramsClosingParen, options[node.type].returnType)
@@ -1219,9 +732,9 @@ export default createRule<RuleOptions, MessageIds>({
       },
 
       IfStatement(node) {
-        addBlocklessNodeIndent(node.consequent)
+        addBlocklessNodeIndent(ctx, node.consequent)
         if (node.alternate)
-          addBlocklessNodeIndent(node.alternate)
+          addBlocklessNodeIndent(ctx, node.alternate)
       },
 
       /**
@@ -1275,7 +788,7 @@ export default createRule<RuleOptions, MessageIds>({
           const openingCurly = sourceCode.getFirstToken(node, isOpeningBraceToken)!
           const closingCurly = sourceCode.getTokenBefore(node.source, isClosingBraceToken)!
 
-          addElementListIndent(node.specifiers.filter(specifier => specifier.type === 'ImportSpecifier'), openingCurly, closingCurly, options.ImportDeclaration)
+          addElementListIndent(ctx, node.specifiers.filter(specifier => specifier.type === 'ImportSpecifier'), openingCurly, closingCurly, options.ImportDeclaration)
         }
 
         const fromToken = node.specifiers.length
@@ -1297,7 +810,7 @@ export default createRule<RuleOptions, MessageIds>({
             offsets.setDesiredOffsets([withToken.range[0], lastToken.range[1]], sourceCode.getFirstToken(node)!, 1)
           }
 
-          addElementListIndent(node.attributes, openingCurly, closingCurly, 1)
+          addElementListIndent(ctx, node.attributes, openingCurly, closingCurly, 1)
         }
       },
 
@@ -1309,15 +822,15 @@ export default createRule<RuleOptions, MessageIds>({
         parameterParens.add(closingParen)
         offsets.setDesiredOffset(openingParen, sourceCode.getTokenBefore(openingParen)!, 0)
 
-        addElementListIndent([node.source], openingParen, closingParen, options.CallExpression.arguments)
+        addElementListIndent(ctx, [node.source], openingParen, closingParen, options.CallExpression.arguments)
       },
 
       MemberExpression(node) {
-        checkMemberExpression(node, node.object, node.property, node.computed)
+        checkMemberExpression(ctx, node, node.object, node.property, node.computed)
       },
 
       MetaProperty(node) {
-        checkMemberExpression(node, node.meta, node.property)
+        checkMemberExpression(ctx, node, node.meta, node.property)
       },
 
       NewExpression(node) {
@@ -1325,7 +838,7 @@ export default createRule<RuleOptions, MessageIds>({
         if (node.arguments.length > 0
           || isClosingParenToken(sourceCode.getLastToken(node)!)
           && isOpeningParenToken(sourceCode.getLastToken(node, 1)!)) {
-          addFunctionCallIndent(node)
+          addFunctionCallIndent(ctx, node)
         }
       },
 
@@ -1337,16 +850,16 @@ export default createRule<RuleOptions, MessageIds>({
         }
       },
 
-      'PropertyDefinition': checkClassProperty,
-      'AccessorProperty': checkClassProperty,
-      'TSAbstractPropertyDefinition': checkClassProperty,
-      'TSAbstractAccessorProperty': checkClassProperty,
+      'PropertyDefinition': node => checkClassProperty(ctx, node),
+      'AccessorProperty': node => checkClassProperty(ctx, node),
+      'TSAbstractPropertyDefinition': node => checkClassProperty(ctx, node),
+      'TSAbstractAccessorProperty': node => checkClassProperty(ctx, node),
 
       StaticBlock(node) {
         const openingCurly = sourceCode.getFirstToken(node, { skip: 1 })! // skip the `static` token
         const closingCurly = sourceCode.getLastToken(node)!
 
-        addElementListIndent(node.body, openingCurly, closingCurly, options.StaticBlock.body)
+        addElementListIndent(ctx, node.body, openingCurly, closingCurly, options.StaticBlock.body)
       },
 
       SwitchStatement(node) {
@@ -1436,6 +949,7 @@ export default createRule<RuleOptions, MessageIds>({
         if (hasDeclaratorOnNewLine) {
           if (alignFirstVariable) {
             addElementListIndent(
+              ctx,
               node.declarations,
               firstToken,
               lastToken,
@@ -1481,7 +995,7 @@ export default createRule<RuleOptions, MessageIds>({
       VariableDeclarator(node) {
         if (node.init) {
           const operator = sourceCode.getTokenBefore(node.init, isNotOpeningParenToken)!
-          checkAssignmentOperator(operator)
+          checkAssignmentOperator(ctx, operator)
         }
 
         const lastToken = sourceCode.getLastToken(node)!
@@ -1533,12 +1047,13 @@ export default createRule<RuleOptions, MessageIds>({
           return
 
         const operator = sourceCode.getFirstTokenBetween(node.name, node.value, isEqToken)!
-        checkAssignmentOperator(operator)
+        checkAssignmentOperator(ctx, operator)
       },
 
       JSXElement(node) {
         if (node.closingElement) {
           addElementListIndent(
+            ctx,
             node.children,
             sourceCode.getFirstToken(node.openingElement)!,
             sourceCode.getFirstToken(node.closingElement)!,
@@ -1559,7 +1074,7 @@ export default createRule<RuleOptions, MessageIds>({
           closingToken = sourceCode.getLastToken(node)!
         }
         offsets.setDesiredOffsets(node.name.range, firstToken, 0)
-        addElementListIndent(node.attributes, firstToken, closingToken, 1)
+        addElementListIndent(ctx, node.attributes, firstToken, closingToken, 1)
       },
 
       JSXClosingElement(node) {
@@ -1572,7 +1087,7 @@ export default createRule<RuleOptions, MessageIds>({
         const firstOpeningToken = sourceCode.getFirstToken(node.openingFragment)!
         const firstClosingToken = sourceCode.getFirstToken(node.closingFragment)!
 
-        addElementListIndent(node.children, firstOpeningToken, firstClosingToken, 1)
+        addElementListIndent(ctx, node.children, firstOpeningToken, firstClosingToken, 1)
       },
 
       JSXOpeningFragment(node) {
@@ -1620,7 +1135,7 @@ export default createRule<RuleOptions, MessageIds>({
       },
 
       JSXMemberExpression(node) {
-        checkMemberExpression(node, node.object, node.property)
+        checkMemberExpression(ctx, node, node.object, node.property)
       },
 
       TSTypeAnnotation(node) {
@@ -1635,27 +1150,27 @@ export default createRule<RuleOptions, MessageIds>({
 
       TSTypeAliasDeclaration(node) {
         const operator = sourceCode.getTokenBefore(node.typeAnnotation, isNotOpeningParenToken)!
-        checkAssignmentOperator(operator)
+        checkAssignmentOperator(ctx, operator)
 
         const lastToken = sourceCode.getLastToken(node)!
         if (isSemicolonToken(lastToken))
           offsets.ignoreToken(lastToken)
       },
 
-      'TSTupleType': checkArrayLikeNode,
+      'TSTupleType': node => checkArrayLikeNode(ctx, node),
 
-      'TSEnumBody': checkBlockLikeNode,
+      'TSEnumBody': node => checkBlockLikeNode(ctx, node),
 
       TSEnumMember(node) {
         if (!node.initializer)
           return
 
         const operator = sourceCode.getTokenBefore(node.initializer, isEqToken)!
-        checkAssignmentOperator(operator)
+        checkAssignmentOperator(ctx, operator)
       },
 
       TSTypeLiteral(node) {
-        checkObjectLikeNode(node, node.members)
+        checkObjectLikeNode(ctx, node, node.members)
       },
 
       TSMappedType(node) {
@@ -1667,19 +1182,19 @@ export default createRule<RuleOptions, MessageIds>({
       },
 
       TSAsExpression(node) {
-        checkOperatorToken(node.expression, node.typeAnnotation, 'as')
+        checkOperatorToken(ctx, node.expression, node.typeAnnotation, 'as')
       },
 
       // TODO: TSSatisfiesExpression
 
       TSConditionalType(node) {
-        checkConditionalNode(node, node.extendsType, node.trueType, node.falseType)
+        checkConditionalNode(ctx, node, node.extendsType, node.trueType, node.falseType)
       },
 
       TSImportEqualsDeclaration(node) {
         if (node.moduleReference) {
           const operator = sourceCode.getTokenBefore(node.moduleReference, isEqToken)!
-          checkAssignmentOperator(operator)
+          checkAssignmentOperator(ctx, operator)
         }
 
         const lastToken = sourceCode.getLastToken(node)!
@@ -1688,20 +1203,20 @@ export default createRule<RuleOptions, MessageIds>({
       },
 
       TSIndexedAccessType(node) {
-        checkMemberExpression(node, node.objectType, node.indexType, true)
+        checkMemberExpression(ctx, node, node.objectType, node.indexType, true)
       },
 
-      'TSInterfaceBody': checkBlockLikeNode,
+      'TSInterfaceBody': node => checkBlockLikeNode(ctx, node),
 
       TSInterfaceDeclaration(node) {
         if (node.extends.length === 0)
           return
 
-        checkHeritages(node, node.extends)
+        checkHeritages(ctx, node, node.extends)
       },
 
       TSQualifiedName(node) {
-        checkMemberExpression(node, node.left, node.right)
+        checkMemberExpression(ctx, node, node.left, node.right)
       },
 
       TSTypeParameter(node) {
@@ -1709,7 +1224,7 @@ export default createRule<RuleOptions, MessageIds>({
           return
 
         const operator = sourceCode.getTokenBefore(node.default, isEqToken)!
-        checkAssignmentOperator(operator)
+        checkAssignmentOperator(ctx, operator)
       },
 
       TSTypeParameterDeclaration(node) {
@@ -1719,7 +1234,7 @@ export default createRule<RuleOptions, MessageIds>({
         const firstToken = sourceCode.getFirstToken(node)!
         const closingToken = sourceCode.getLastToken(node)!
 
-        addElementListIndent(node.params, firstToken, closingToken, 1)
+        addElementListIndent(ctx, node.params, firstToken, closingToken, 1)
       },
 
       TSTypeParameterInstantiation(node) {
@@ -1729,10 +1244,10 @@ export default createRule<RuleOptions, MessageIds>({
         const firstToken = sourceCode.getFirstToken(node)!
         const closingToken = sourceCode.getLastToken(node)!
 
-        addElementListIndent(node.params, firstToken, closingToken, 1)
+        addElementListIndent(ctx, node.params, firstToken, closingToken, 1)
       },
 
-      'TSModuleBlock': checkBlockLikeNode,
+      'TSModuleBlock': node => checkBlockLikeNode(ctx, node),
 
       '*': function (node: ASTNode) {
         const firstToken = sourceCode.getFirstToken(node)
@@ -1820,9 +1335,9 @@ export default createRule<RuleOptions, MessageIds>({
         }
 
         // Update the offsets for ignored nodes to prevent their child tokens from being reported.
-        ignoredNodes.forEach(ignoreNode)
+        ignoredNodes.forEach(node => ignoreNode(ctx, node))
 
-        addParensIndent(sourceCode.ast.tokens)
+        addParensIndent(ctx, sourceCode.ast.tokens)
 
         /**
          * Create a Map from (tokenOrComment) => (precedingToken).
